Move ESLint config to a common module
